+++ src/pages/results/results-overview.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { ResultsOverviewPage } from './results-overview';
+import { SuperTabsModule } from 'ionic2-super-tabs';
 
 @NgModule({
   declarations: [
@@ -8,6 +9,7 @@ import { ResultsOverviewPage } from './results-overview';
   ],
   imports: [
     IonicPageModule.forChild(ResultsOverviewPage),
+    SuperTabsModule
   ],
 })
 export class ResultsOverviewPageModule {}