+++ src/pages/survey/surveys-overview/surveys-overview.module.ts
@@ -2,12 +2,14 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { SurveysOverviewPage } from './surveys-overview';
 
+
 @NgModule({
   declarations: [
     SurveysOverviewPage,
   ],
   imports: [
     IonicPageModule.forChild(SurveysOverviewPage),
+    
   ],
 })
 export class SurveysOverviewPageModule {}