+++ changelog.md
@@ -1,8 +1,13 @@
 #ChangeLog
+##0.1.2
 
 ##0.1.1
 - Farmer enrollment
 - Experiment detail page (click (i) or on text in available experiments)
+- Surveys linked to experiments
+- GroPlus Survey (part 1)
+- Various database improvements and changes
+- Results page placeholder
 
 ##0.1.0
 First alpha release, includes: